backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
pydantic-settings==2.3.0
loguru==0.7.0
google-generativeai==0.7.5
httpx==0.28.0
pytest==8.1.1
openai==1.55.3
